feat/modulo 3

+++ src/modules/projects/projects.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   Get,
   HttpCode,
+  HttpException,
   HttpStatus,
   Param,
   ParseUUIDPipe,
@@ -29,8 +30,14 @@ export class ProjectsController {
 
   @Get(':id')
   @ApiResponse({ type: ProjectsListItemDTO })
-  findOne(@Param('id', ParseUUIDPipe) id: string) {
-    return this.projectsService.findById(id)
+  async findOne(@Param('id', ParseUUIDPipe) id: string) {
+    const project = await this.projectsService.findById(id)
+
+    if (!project) {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND)
+    }
+
+    return project
   }
 
   @Post()
@@ -41,13 +48,25 @@ export class ProjectsController {
 
   @Put(':id')
   @ApiResponse({ type: ProjectsListItemDTO })
-  update(@Param('id', ParseUUIDPipe) id: string, @Body() data: ProjectsRequestDto) {
+  async update(@Param('id', ParseUUIDPipe) id: string, @Body() data: ProjectsRequestDto) {
+    const project = await this.projectsService.findById(id)
+
+    if (!project) {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND)
+    }
+
     return this.projectsService.update(id, data)
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  remove(@Param('id', ParseUUIDPipe) id: string) {
+  async remove(@Param('id', ParseUUIDPipe) id: string) {
+    const project = await this.projectsService.findById(id)
+
+    if (!project) {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND)
+    }
+
     return this.projectsService.remove(id)
   }
 }
